Pivot to SBT for build tool.
This might be contentious to non-Scala devs, but the ease of enabling Travis
CI publishing to Maven Central via the sbt-ci-release plugin is very valuable.
I briefly tried to get Gradle working and it was painful; several plugins
have to be carefully orchestrated, and multiple files need to be encrypted
and checked in. sbt-ci-release makes it all a lot easier.

See https://github.com/jameskoch/jvm-repr/commit/dfab9c70105c695673321f2cf641d8ea79c541e1
for notes and progress on the Gradle stuff.